+++ request/src/test/java/karate/request/client/client.java
@@ -1,10 +1,10 @@
-package karate.request;
+package karate.request.client;
 
 import com.intuit.karate.junit5.Karate;
 
 public class client {
     @Karate.Test
     Karate testSample(){
-        return Karate.run("classpath:karate/request/client.feature").relativeTo(getClass());
+        return Karate.run("classpath:karate/request/client_Consult_Get/client.feature").relativeTo(getClass());
     }
 }
